+++ Client/Pangu/Assets/Code/Pangu/Runtime/3rd/com.tuyoogame.yooasset@2.0.3-preview/Runtime/Extensions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0c0e350e7a5e24746b95abdd8c49d63c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
