+++ Client/Pangu/Assets/Code/Pangu/Runtime/3rd/com.tuyoogame.yooasset@2.0.3-preview/Runtime/Extensions/YooAssetsExtension.AssetDownLoad.cs
@@ -1,0 +1,91 @@
+namespace YooAsset
+{
+    public static partial class YooAssets
+    {
+        #region 资源下载
+        /// <summary>
+        /// 创建资源下载器，用于下载当前资源版本所有的资源包文件
+        /// </summary>
+        /// <param name="downloadingMaxNumber">同时下载的最大文件数</param>
+        /// <param name="failedTryAgain">下载失败的重试次数</param>
+        public static ResourceDownloaderOperation CreateResourceDownloader(int downloadingMaxNumber, int failedTryAgain)
+        {
+            DebugCheckDefaultPackageValid();
+            return _defaultPackage.CreateResourceDownloader(downloadingMaxNumber, failedTryAgain);
+        }
+
+        /// <summary>
+        /// 创建资源下载器，用于下载指定的资源标签关联的资源包文件
+        /// </summary>
+        /// <param name="tag">资源标签</param>
+        /// <param name="downloadingMaxNumber">同时下载的最大文件数</param>
+        /// <param name="failedTryAgain">下载失败的重试次数</param>
+        public static ResourceDownloaderOperation CreateResourceDownloader(string tag, int downloadingMaxNumber, int failedTryAgain)
+        {
+            DebugCheckDefaultPackageValid();
+            return _defaultPackage.CreateResourceDownloader(new string[] { tag }, downloadingMaxNumber, failedTryAgain);
+        }
+
+        /// <summary>
+        /// 创建资源下载器，用于下载指定的资源标签列表关联的资源包文件
+        /// </summary>
+        /// <param name="tags">资源标签列表</param>
+        /// <param name="downloadingMaxNumber">同时下载的最大文件数</param>
+        /// <param name="failedTryAgain">下载失败的重试次数</param>
+        public static ResourceDownloaderOperation CreateResourceDownloader(string[] tags, int downloadingMaxNumber, int failedTryAgain)
+        {
+            DebugCheckDefaultPackageValid();
+            return _defaultPackage.CreateResourceDownloader(tags, downloadingMaxNumber, failedTryAgain);
+        }
+
+        /// <summary>
+        /// 创建资源下载器，用于下载指定的资源依赖的资源包文件
+        /// </summary>
+        /// <param name="location">资源定位地址</param>
+        /// <param name="downloadingMaxNumber">同时下载的最大文件数</param>
+        /// <param name="failedTryAgain">下载失败的重试次数</param>
+        public static ResourceDownloaderOperation CreateBundleDownloader(string location, int downloadingMaxNumber, int failedTryAgain)
+        {
+            DebugCheckDefaultPackageValid();
+            return _defaultPackage.CreateBundleDownloader(location, downloadingMaxNumber, failedTryAgain);
+        }
+
+        /// <summary>
+        /// 创建资源下载器，用于下载指定的资源列表依赖的资源包文件
+        /// </summary>
+        /// <param name="locations">资源定位地址列表</param>
+        /// <param name="downloadingMaxNumber">同时下载的最大文件数</param>
+        /// <param name="failedTryAgain">下载失败的重试次数</param>
+        public static ResourceDownloaderOperation CreateBundleDownloader(string[] locations, int downloadingMaxNumber, int failedTryAgain)
+        {
+            DebugCheckDefaultPackageValid();
+            return _defaultPackage.CreateBundleDownloader(locations, downloadingMaxNumber, failedTryAgain);
+        }
+
+        /// <summary>
+        /// 创建资源下载器，用于下载指定的资源依赖的资源包文件
+        /// </summary>
+        /// <param name="assetInfo">资源信息</param>
+        /// <param name="downloadingMaxNumber">同时下载的最大文件数</param>
+        /// <param name="failedTryAgain">下载失败的重试次数</param>
+        public static ResourceDownloaderOperation CreateBundleDownloader(AssetInfo assetInfo, int downloadingMaxNumber, int failedTryAgain)
+        {
+            DebugCheckDefaultPackageValid();
+            return _defaultPackage.CreateBundleDownloader(assetInfo, downloadingMaxNumber, failedTryAgain);
+        }
+
+        /// <summary>
+        /// 创建资源下载器，用于下载指定的资源列表依赖的资源包文件
+        /// </summary>
+        /// <param name="assetInfos">资源信息列表</param>
+        /// <param name="downloadingMaxNumber">同时下载的最大文件数</param>
+        /// <param name="failedTryAgain">下载失败的重试次数</param>
+        public static ResourceDownloaderOperation CreateBundleDownloader(AssetInfo[] assetInfos, int downloadingMaxNumber, int failedTryAgain)
+        {
+            DebugCheckDefaultPackageValid();
+            return _defaultPackage.CreateBundleDownloader(assetInfos, downloadingMaxNumber, failedTryAgain);
+        }
+        #endregion
+    }
+}
+
