+++ Client/Pangu/Assets/Code/Pangu/Runtime/3rd/com.tuyoogame.yooasset@2.0.3-preview/Runtime/Extensions/YooAssetsExtension.Debug.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 70cf1478eb78c984b884fee1fcb15344
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
