+++ Client/Pangu/Assets/Code/Pangu/Runtime/3rd/com.tuyoogame.yooasset@2.0.3-preview/Runtime/Extensions/YooAssetsExtension.SceneLoad.cs
@@ -1,0 +1,39 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+namespace YooAsset
+{
+    public static partial class YooAssets
+    {
+        #region 场景加载
+        /// <summary>
+        /// 异步加载场景
+        /// </summary>
+        /// <param name="location">场景的定位地址</param>
+        /// <param name="sceneMode">场景加载模式</param>
+        /// <param name="suspendLoad">场景加载到90%自动挂起</param>
+        /// <param name="priority">优先级</param>
+        public static SceneHandle LoadSceneAsync(string location, LoadSceneMode sceneMode = LoadSceneMode.Single, bool suspendLoad = false, uint priority = 100)
+        {
+            DebugCheckDefaultPackageValid();
+            return _defaultPackage.LoadSceneAsync(location, sceneMode, suspendLoad, priority);
+        }
+
+        /// <summary>
+        /// 异步加载场景
+        /// </summary>
+        /// <param name="assetInfo">场景的资源信息</param>
+        /// <param name="sceneMode">场景加载模式</param>
+        /// <param name="suspendLoad">场景加载到90%自动挂起</param>
+        /// <param name="priority">优先级</param>
+        public static SceneHandle LoadSceneAsync(AssetInfo assetInfo, LoadSceneMode sceneMode = LoadSceneMode.Single, bool suspendLoad = false, uint priority = 100)
+        {
+            DebugCheckDefaultPackageValid();
+            return _defaultPackage.LoadSceneAsync(assetInfo, sceneMode, suspendLoad, priority);
+        }
+        #endregion
+    }
+}
+
