+++ Client/Pangu/Assets/Code/Pangu/Runtime/3rd/com.tuyoogame.yooasset@2.0.3-preview/Runtime/Extensions/YooAssetsExtension.SceneLoad.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7615c1c9eb5363349a3d99cbe5d3476f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
